+++ CHANGELOG.md
@@ -4,6 +4,7 @@ CHANGELOG
 unreleased
 ----------
 - Hide Diners Club logo when Diners Club may not be supported
+- Add `cardholderName` option to card configuration
 - Use version 3.21.1 of braintree-web
 
 1.5.0